Relocate stock Roda plugins to their own files
